Expose `EditableProps` from slate-react component
Not sure if this is acceptable. I have a usecase where I want to wrap an `Editable` with some custom handling, but pass through any properties that might be specified. 

As an example, consider:

```typescript
export const WrappedEditable: FC<MyProps & EditableProps> = ({
  myProp,
  onKeyDown,
  ...editableProps
}) => {

myOnKeyDown = () => { /* whatever*/}
return <Editable onKeyDown={myOnKeyDown} {...editableProps} />
}
```

In typescript, I needed to copy-paste `EditableProps` into my own code in order to access this. Feel free to close if it's not ok to expose it.